Move Account.kt to commons: replace 5 concrete deps with factory lambdas
- Replace EventProcessor with eventProcessorFactory: (IAccount, ICacheProvider) -> IEventProcessor
- Replace NwcSignerState with nwcSignerStateFactory
- Replace TrustProviderListState with trustProviderListStateFactory
- Replace FeedTopNavFilterState with feedTopNavFilterStateFactory
- Change NWCPaymentFilterAssembler type to INWCPaymentFilterAssembler
- Move Account.kt to commons/src/jvmAndroid/ (uses OtsState which is JVM-only)
- Leave typealias at original amethyst/model/Account.kt location
- Change cache type from LocalCache to ICacheProvider
- Move anti-spam monitoring to AccountCacheState construction site
- Add sendNwcRequest/sendZapPaymentRequestFor/hasWalletConnectSetup to INwcSignerState
- Add observeEvents/observeLatestEvent/getAnyChannel(Event) to ICacheProvider
- Update AccountCacheState and AccountViewModel construction sites
- Update DesktopIAccount with new INwcSignerState methods
